static-cast float and reducing variables
- replaced all C casting floats on window height and width to static_cast floats
- further improved alien and background by reducing variables and loops

+++ SpaceInvaders_2023/Source/Player.cpp
@@ -10,7 +10,7 @@ void Player::UpdateMovement()
     else if (IsKeyDown(KEY_RIGHT)) {
         movement = 1;
     }
-    x_pos = std::clamp(x_pos += speed * movement, PLAYER_RADIUS, GetScreenWidth() - PLAYER_RADIUS);
+    x_pos = std::clamp(x_pos += speed * movement, PLAYER_RADIUS, static_cast<float>(GetScreenWidth()) - PLAYER_RADIUS);
 }
 
 void Player::UpdateAnimation() noexcept
@@ -32,6 +32,6 @@ void Player::Update()
 void Player::Render(Texture2D texture) noexcept
 {
     const Rectangle DT_src = { 0,0,static_cast<float>(texture.width), static_cast<float>(texture.height) };
-    const Rectangle DT_dst = { x_pos, GetScreenHeight() - PLAYER_BASE_HEIGHT, 100, 100 };
+    const Rectangle DT_dst = { x_pos, static_cast<float>(GetScreenHeight()) - PLAYER_BASE_HEIGHT, 100, 100 };
     DrawTexturePro(texture, DT_src, DT_dst, DT_o, 0, WHITE);
 }
